Update authorization and company relationship handling
Made enhancements to the authorization guards and added the capability to link companies to users via the UserCompanyEntity. Also, added the NewsController to handle news requests and introduced checks to ensure the uniqueness of a company's slug and company's relationship in the NewsCategoryEntity. Other minor changes include code cleaning and correcting of some entity relationships.