Improve user module and add company-user contact functionality
Reflecting significant enhancements to the user module, this push incorporates the creation of new mapper files and the addition of user-company contact functionality. Updates to various files and DTOs, along with adjustments to existing mappers and services, align with the goal of improving user and contact management. A significant change includes the ability for users to have multiple contacts and positions defined within different companies.