Update return type in user update method
The return type of the 'update' method in the UserController class has been changed from Promise<IResponseUser> to Promise<ResponseFullUserMapper>. This is part of an enhancement to the user module and aligns with the addition of user-company contact functionality. It improves user and management contact by allowing users to hold multiple contacts and positions within different companies.