Update type of news images and implement image handling
The type for the news images and the portrait image has been updated from string to INewsImage for better structure. Added implementation for image handling in the news service which includes uploading the images and storing the file metadata in the News entity. This also includes changes in the news controller to accommodate these modifications.